Adding SMIA_USER Environment Variable
With many people pushing images with the same name to a single Docker
Machine we're going to end up with clashes as it's impossible to whose
image is whose.  To solve this issue the docker-compole.yml and the pom
now expect the SMIA_USER environment variable to exist which should be
set to the username of the user (or atleast something unique).

Thus an image which was originally named smia/tmx-simple-service becomes
smia/u001807/tmx-simple-service.